Move gitrequestlogic from handlers
