+++ bot/setup.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 import os
 
-TOKEN = ''  # insert your token here
+TOKEN =   # insert your token here
 PROXY = 'https://telegg.ru/orig/bot'  # proxy to connect against telegram ban
